restructure non_local_operator.hpp
- split class and method definition
- add `s_u_operator.cpp`
